+++ src/Aihc/Parser/Internal/Expr.hs
@@ -21,7 +21,7 @@ import Aihc.Parser.Internal.Pattern (apatParser, caseAltPatternParser, patParser
 import Aihc.Parser.Internal.Type (typeAtomParser, typeParser, typeSignatureParser)
 import Aihc.Parser.Lex (LexToken (..), LexTokenKind (..), lexTokenKind, lexTokenSpan, lexTokenText)
 import Aihc.Parser.Syntax
-import Aihc.Parser.Types (ParserErrorComponent (..), TokStream (..), mkFoundToken)
+import Aihc.Parser.Types (ParserErrorComponent (..), mkFoundToken)
 import Control.Monad (guard)
 import Data.Functor (($>))
 import Data.Text (Text)
@@ -158,14 +158,14 @@ multiWayIfExprParser = withSpanAnn (EAnn . mkAnnotation) $ do
   EMultiWayIf <$> braces (MP.some multiWayIfAlternative)
 
 multiWayIfAlternative :: TokParser (GuardedRhs Expr)
-multiWayIfAlternative = withSpan $ do
+multiWayIfAlternative = withSpanAnns $ do
   expectedTok TkReservedPipe
   guards <- layoutSepBy1 (guardQualifierParser RhsArrowCase) (expectedTok TkSpecialComma)
   expectedTok TkReservedRightArrow
   body <- exprParser
-  pure $ \span' ->
+  pure $ \anns ->
     GuardedRhs
-      { guardedRhsAnns = [mkAnnotation span'],
+      { guardedRhsAnns = anns,
         guardedRhsGuards = guards,
         guardedRhsBody = body
       }
@@ -388,10 +388,8 @@ appExprParserWith atomParser = do
     [] -> pure first
     _ -> do
       endInput <- MP.getInput
-      let startSpan = inputStartSpan startInput
-          endSpan = maybe noSourceSpan lexTokenSpan (tokStreamPrevToken endInput)
-          appSpan = mergeSourceSpans startSpan endSpan
-      pure (EAnn (mkAnnotation appSpan) (foldl applyArg first rest))
+      let app = foldl applyArg first rest
+      pure (maybe app (\sp -> EAnn (mkAnnotation sp) app) (consumedSpan startInput endInput))
   where
     appArg :: TokParser (Either Type Expr)
     appArg = (Left <$> typeAppArg) <|> (Right <$> appExprArgParser)
@@ -465,11 +463,11 @@ atomOrRecordExprParserWith atomContext =
               fieldName <- recordFieldNameParser
               applyRecordSuffixes (EGetField e fieldName)
 
-    normalizeField :: (Name, Maybe Expr, SourceSpan) -> RecordField Expr
-    normalizeField (fieldName, mExpr, sp) =
+    normalizeField :: (Name, Maybe Expr, [Annotation]) -> RecordField Expr
+    normalizeField (fieldName, mExpr, anns) =
       case mExpr of
         Just expr' -> RecordField fieldName expr' False
-        Nothing -> RecordField fieldName (EAnn (mkAnnotation sp) (EVar fieldName)) True
+        Nothing -> RecordField fieldName (foldr EAnn (EVar fieldName) anns) True
 
     recordDotMayFollow :: Expr -> Bool
     recordDotMayFollow expr =
@@ -481,13 +479,13 @@ atomOrRecordExprParserWith atomContext =
         _ -> True
 
 -- | Parse record braces: { field = value, field2 = value2, ... }
-recordBracesParser :: TokParser ([(Name, Maybe Expr, SourceSpan)], Bool)
+recordBracesParser :: TokParser ([(Name, Maybe Expr, [Annotation])], Bool)
 recordBracesParser =
   braces $
     recordFieldsWithWildcardsParser (layoutSepEndBy recordFieldBindingParser (expectedTok TkSpecialComma))
 
-recordFieldBindingParser :: TokParser (Name, Maybe Expr, SourceSpan)
-recordFieldBindingParser = withSpan $ do
+recordFieldBindingParser :: TokParser (Name, Maybe Expr, [Annotation])
+recordFieldBindingParser = withSpanAnns $ do
   fieldName <- recordFieldNameParser
   mAssign <- MP.optional (expectedTok TkReservedEquals *> exprParser)
   pure (fieldName,mAssign,)
@@ -697,31 +695,31 @@ rhsParserWithBodyParser arrowKind bodyParser = do
           }
 
 unguardedRhsParserWithBodyParser :: RhsArrowKind -> TokParser body -> TokParser (Rhs body)
-unguardedRhsParserWithBodyParser arrowKind bodyParser = withSpan $ do
+unguardedRhsParserWithBodyParser arrowKind bodyParser = withSpanAnns $ do
   rhsArrowTok arrowKind
   body <- region (rhsContextText arrowKind) bodyParser
   whereDecls <- MP.optional whereClauseParser
-  pure (\span' -> UnguardedRhs [mkAnnotation span'] body whereDecls)
+  pure (\anns -> UnguardedRhs anns body whereDecls)
 
 rhsContextText :: RhsArrowKind -> Text
 rhsContextText RhsArrowCase = "while parsing case alternative right-hand side"
 rhsContextText RhsArrowEquation = "while parsing equation right-hand side"
 
 guardedRhssParserWithBodyParser :: RhsArrowKind -> TokParser body -> TokParser (Rhs body)
-guardedRhssParserWithBodyParser arrowKind bodyParser = withSpan $ do
+guardedRhssParserWithBodyParser arrowKind bodyParser = withSpanAnns $ do
   grhss <- MP.some (guardedRhsParserWithBodyParser arrowKind bodyParser)
   whereDecls <- MP.optional whereClauseParser
-  pure (\span' -> GuardedRhss [mkAnnotation span'] grhss whereDecls)
+  pure (\anns -> GuardedRhss anns grhss whereDecls)
 
 guardedRhsParserWithBodyParser :: RhsArrowKind -> TokParser body -> TokParser (GuardedRhs body)
-guardedRhsParserWithBodyParser arrowKind bodyParser = withSpan $ do
+guardedRhsParserWithBodyParser arrowKind bodyParser = withSpanAnns $ do
   expectedTok TkReservedPipe
   guards <- layoutSepBy1 (guardQualifierParser arrowKind) (expectedTok TkSpecialComma)
   rhsArrowTok arrowKind
   body <- bodyParser
-  pure $ \span' ->
+  pure $ \anns ->
     GuardedRhs
-      { guardedRhsAnns = [mkAnnotation span'],
+      { guardedRhsAnns = anns,
         guardedRhsGuards = guards,
         guardedRhsBody = body
       }
@@ -762,23 +760,23 @@ guardExprParser RhsArrowEquation = exprParserWithTypeSigParser typeParser
 guardExprParser RhsArrowCase = exprCoreParserWithoutTypeSig
 
 caseAltParser :: TokParser (CaseAlt Expr)
-caseAltParser = withSpan $ do
+caseAltParser = withSpanAnns $ do
   pat <- region "while parsing case alternative" caseAltPatternParser
   rhs <- region "while parsing case alternative" rhsParser
-  pure $ \span' ->
+  pure $ \anns ->
     CaseAlt
-      { caseAltAnns = [mkAnnotation span'],
+      { caseAltAnns = anns,
         caseAltPattern = pat,
         caseAltRhs = rhs
       }
 
 lambdaCaseAltParser :: TokParser LambdaCaseAlt
-lambdaCaseAltParser = withSpan $ do
+lambdaCaseAltParser = withSpanAnns $ do
   pats <- region "while parsing lambda-cases alternative" (MP.many apatParser)
   rhs <- region "while parsing lambda-cases alternative" rhsParser
-  pure $ \span' ->
+  pure $ \anns ->
     LambdaCaseAlt
-      { lambdaCaseAltAnns = [mkAnnotation span'],
+      { lambdaCaseAltAnns = anns,
         lambdaCaseAltPats = pats,
         lambdaCaseAltRhs = rhs
       }
@@ -1135,23 +1133,23 @@ compTransformLambdaExprParser = withSpanAnn (EAnn . mkAnnotation) $ do
       pure (ELambdaPats pats body)
 
 compTransformCaseAltParser :: TokParser (CaseAlt Expr)
-compTransformCaseAltParser = withSpan $ do
+compTransformCaseAltParser = withSpanAnns $ do
   pat <- region "while parsing case alternative" caseAltPatternParser
   rhs <- region "while parsing case alternative" (caseRhsParserWithBodyParser compTransformExprParser)
-  pure $ \span' ->
+  pure $ \anns ->
     CaseAlt
-      { caseAltAnns = [mkAnnotation span'],
+      { caseAltAnns = anns,
         caseAltPattern = pat,
         caseAltRhs = rhs
       }
 
 compTransformLambdaCaseAltParser :: TokParser LambdaCaseAlt
-compTransformLambdaCaseAltParser = withSpan $ do
+compTransformLambdaCaseAltParser = withSpanAnns $ do
   pats <- region "while parsing lambda-cases alternative" (MP.many apatParser)
   rhs <- region "while parsing lambda-cases alternative" (caseRhsParserWithBodyParser compTransformExprParser)
-  pure $ \span' ->
+  pure $ \anns ->
     LambdaCaseAlt
-      { lambdaCaseAltAnns = [mkAnnotation span'],
+      { lambdaCaseAltAnns = anns,
         lambdaCaseAltPats = pats,
         lambdaCaseAltRhs = rhs
       }