+++ src/Aihc/Parser/Internal/Pattern.hs
@@ -102,14 +102,8 @@ lpatParser =
 buildPatternApp :: Pattern -> Pattern -> Pattern
 buildPatternApp lhs rhs =
   case peelPatternAnn lhs of
-    PCon name typeArgs args ->
-      PAnn
-        (mkAnnotation NoSourceSpan)
-        (PCon name typeArgs (args <> [rhs]))
-    PBuiltinCon con typeArgs args ->
-      PAnn
-        (mkAnnotation NoSourceSpan)
-        (PBuiltinCon con typeArgs (args <> [rhs]))
+    PCon name typeArgs args -> PCon name typeArgs (args <> [rhs])
+    PBuiltinCon con typeArgs args -> PBuiltinCon con typeArgs (args <> [rhs])
     _ -> lhs
 
 -- | Parse an atomic pattern (@apat@ in the Haskell Report).
@@ -280,10 +274,10 @@ thSplicePatternParser = withSpanAnn (PAnn . mkAnnotation) $ do
 
 visibleTypeBinderCoreParser :: TokParser TyVarBinder
 visibleTypeBinderCoreParser =
-  withSpan $
+  withSpanAnns $
     ( do
         ident <- tyVarNameParser
-        pure (\span' -> TyVarBinder [mkAnnotation span'] ident Nothing TyVarBSpecified TyVarBInvisible)
+        pure (\anns -> TyVarBinder anns ident Nothing TyVarBSpecified TyVarBInvisible)
     )
       <|> ( do
               expectedTok TkSpecialLParen
@@ -291,7 +285,7 @@ visibleTypeBinderCoreParser =
               expectedTok TkReservedDoubleColon
               kind <- typeParser
               expectedTok TkSpecialRParen
-              pure (\span' -> TyVarBinder [mkAnnotation span'] ident (Just kind) TyVarBSpecified TyVarBInvisible)
+              pure (\anns -> TyVarBinder anns ident (Just kind) TyVarBSpecified TyVarBInvisible)
           )
 
 varOrConPatternParser :: TokParser Pattern
@@ -304,8 +298,8 @@ varOrConPatternParser = do
       | isConLikeName name && lexTokenKind nextTok == TkSpecialLBrace -> do
           (fields, hasWildcard) <- braces recordPatternFieldListParser
           endInput <- MP.getInput
-          let endSpan = maybe noSourceSpan lexTokenSpan (tokStreamPrevToken endInput)
-              recordSpan = mergeSourceSpans (lexTokenSpan tok) endSpan
+          let startSpan = lexTokenSpan tok
+              recordSpan = maybe startSpan (mergeSourceSpans startSpan . lexTokenSpan) (tokStreamPrevToken endInput)
           pure (PAnn (mkAnnotation recordSpan) (PRecord name fields hasWildcard))
     _ ->
       pure $